+++ packages/web-core/src/common/data_structures.ts
@@ -108,6 +108,7 @@ export interface ClientConfiguration {
   capabilities?: Capabilities;
   audioPlayback?: AudioPlaybackConfig;
   history?: HistoryConfig;
+  validateData?: boolean;
 }
 
 export interface InternalClientConfiguration {
@@ -116,6 +117,7 @@ export interface InternalClientConfiguration {
   capabilities: CapabilitiesConfiguration;
   audioPlayback?: AudioPlaybackConfig;
   history?: HistoryConfig;
+  validateData?: boolean;
 }
 
 export interface CancelResponses {